Сделана минимальная рабочая версия приложения

+++ app/src/main/java/com/example/vk_currency_converter/data/dto/ResponseExchangeRates.kt
@@ -1,3 +1,5 @@
 package com.example.vk_currency_converter.data.dto
 
-data class ResponseExchangeRates(val rates: HashMap<String, Double>?) : Response()
+import java.math.BigDecimal
+
+data class ResponseExchangeRates(val rates: HashMap<String, BigDecimal>?) : Response()
